requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
asttokens==2.2.1
attrs==25.4.0
autopep8==2.0.2
backcall==0.2.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==1.3.7
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.3.1
distlib==0.3.7
distro==1.9.0
Django==5.0
dss-python==0.15.7
dss-python-backend==0.14.5
durationpy==0.10
et_xmlfile==2.0.0
executing==1.2.0
fastapi==0.123.9
fastembed==0.7.4
filelock==3.20.0
Flask==2.2.2
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
google-play-scraper==1.2.3
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==0.37.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.2.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.22.0
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==0.3.25
langchain-chroma==1.1.0
langchain-community==0.3.24
langchain-core==1.2.0
langchain-experimental==0.3.4
langchain-groq==0.3.2
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.4.59
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.13.0
nest-asyncio==1.5.6
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
opencv-python==4.12.0.88
opendssdirect.py==0.9.4
openpyxl==3.1.5
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-asgi==0.60b0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opentelemetry-util-http==0.60b0
orjson==3.11.4
overrides==7.7.0
packaging==23.2
pandas==2.3.3
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.1.0
plotly==5.14.1
plotly-express==0.4.1
posthog==5.4.0
prompt-toolkit==3.0.38
propcache==0.4.1
protobuf==6.33.1
psutil==5.9.5
pulsar-client==3.8.0
pure-eval==0.2.2
py-dss-interface==2.2.1
py-dss-toolkit==0.5.0
py_rust_stemmers==0.1.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.10.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.6
pyparsing==3.2.5
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==306
PyYAML==6.0.3
pyzmq==25.0.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.10.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.50.0
statsmodels==0.13.5
streamlit==1.52.1
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.9.0
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
validators==0.35.0
virtualenv==20.25.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.6
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.0.1
win32_setctime==1.2.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
onnxruntime==1.18.0
tiktoken==0.7.0
numpy==1.26.4
python-pptx
